% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{encode}
\alias{encode}
\title{Encode an R object into a CBOR byte vector}
\usage{
encode(x)
}
\arguments{
\item{x}{The R object to encode.}
}
\value{
A \code{raw} vector containing the CBOR representation of the object.
}
\description{
This function takes an R object and serializes it into a \code{raw} vector
representing the object in CBOR format. It supports most common R data
types, including atomic vectors, lists, and \code{NULL}.
}
\examples{
# Encode a simple integer
raw_bytes <- encode(123L)

# Encode a named list
my_list <- list(a = 1, b = "hello", c = c(TRUE, FALSE, NA))
encoded_list <- encode(my_list)

# The output is a raw vector
class(encoded_list)

}
