% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbor.r
\name{rcbor_hello}
\alias{rcbor_hello}
\title{A simple hello world from Rust}
\usage{
rcbor_hello()
}
\value{
A string "Hello world!".
}
\description{
This is a test function to ensure the Rust library is loaded
and callable from R.
}
